rolled back info dict flattening in HabGymWrapper; rolled back https://github.com/facebookresearch/habitat-lab/pull/1060; moved viz functions to visualizations utils
