+++ examples/vln_reference_path_follower_example.py
@@ -22,7 +22,7 @@ from habitat.config.default_structured_configs import (
 from habitat.sims.habitat_simulator.actions import HabitatSimActions
 from habitat.tasks.nav.shortest_path_follower import ShortestPathFollower
 from habitat.utils.visualizations.utils import (
-    append_text_to_image,
+    append_text_underneath_image,
     images_to_video,
 )
 
@@ -35,7 +35,7 @@ def save_map(observations, info, images):
     im = observations["rgb"]
     top_down_map = draw_top_down_map(info, im.shape[0])
     output_im = np.concatenate((im, top_down_map), axis=1)
-    output_im = append_text_to_image(
+    output_im = append_text_underneath_image(
         output_im, observations["instruction"]["text"]
     )
     images.append(output_im)
